Distribute multiparser code
